# SMS-Onboarding-Unified/api-specification.yaml
openapi: 3.0.3
info:
  title: SMS Onboarding Portal API
  description: |
    RESTful API for the SMS Onboarding Portal that manages vessel onboarding, equipment documentation,
    critical parts management, and integration with the maintenance portal.
  version: 1.0.0
  contact:
    name: SMS Support Team
    email: support@sms-portal.com
  license:
    name: Proprietary
servers:
  - url: https://api.sms-onboarding.com/v1
    description: Production server
  - url: https://staging-api.sms-onboarding.com/v1
    description: Staging server
  - url: http://localhost:4000/v1
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /vessels/{vesselId}/equipment:
    get:
      tags:
        - Equipment
      summary: List equipment for a vessel
      operationId: listEquipment
      parameters:
        - $ref: '#/components/parameters/VesselId'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/SortQuery'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/EquipmentType'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/EquipmentStatus'
        - name: criticalLevel
          in: query
          schema:
            $ref: '#/components/schemas/CriticalLevel'
        - name: location
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Equipment list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Equipment
      summary: Create new equipment
      operationId: createEquipment
      parameters:
        - $ref: '#/components/parameters/VesselId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEquipmentRequest'
      responses:
        '201':
          description: Equipment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /equipment/{equipmentId}:
    get:
      tags:
        - Equipment
      summary: Get equipment by ID
      operationId: getEquipment
      parameters:
        - $ref: '#/components/parameters/EquipmentId'
      responses:
        '200':
          description: Equipment retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Equipment
      summary: Update equipment
      operationId: updateEquipment
      parameters:
        - $ref: '#/components/parameters/EquipmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEquipmentRequest'
      responses:
        '200':
          description: Equipment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Equipment
      summary: Delete equipment
      operationId: deleteEquipment
      parameters:
        - $ref: '#/components/parameters/EquipmentId'
      responses:
        '204':
          description: Equipment deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /batch/equipment:
    post:
      tags:
        - Equipment
      summary: Batch create/update equipment
      operationId: batchEquipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - operations
              properties:
                operations:
                  type: array
                  maxItems: 100
                  items:
                    type: object
                    required:
                      - action
                      - data
                    properties:
                      action:
                        type: string
                        enum: [create, update, delete]
                      data:
                        oneOf:
                          - $ref: '#/components/schemas/CreateEquipmentRequest'
                          - $ref: '#/components/schemas/UpdateEquipmentRequest'
                      equipmentId:
                        type: string
                        format: uuid
                        description: Required for update and delete actions
      responses:
        '200':
          description: Batch operation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchOperationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    webhookAuth:
      type: apiKey
      in: header
      name: X-Webhook-Signature
  parameters:
...
    VesselId:
      name: vesselId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EquipmentId:
      name: equipmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageQuery:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitQuery:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    SearchQuery:
      name: search
      in: query
      schema:
        type: string
    SortQuery:
      name: sort
      in: query
      schema:
        type: string
        pattern: '^[a-zA-Z_]+:(asc|desc)$'
  schemas:
...
    CreateEquipmentRequest:
      type: object
      required:
        - name
        - type
        - manufacturer
        - model
        - location
        - criticalLevel
      properties:
        name:
          type: string
        type:
          $ref: '#/components/schemas/EquipmentType'
        manufacturer:
          type: string
        model:
          type: string
        serialNumber:
          type: string
        location:
          type: string
        installationDate:
          type: string
          format: date
        criticalLevel:
          $ref: '#/components/schemas/CriticalLevel'
        status:
          $ref: '#/components/schemas/EquipmentStatus'
        technicalSpecs:
          type: object
          additionalProperties: true
    UpdateEquipmentRequest:
      type: object
      properties:
        name:
          type: string
        type:
          $ref: '#/components/schemas/EquipmentType'
        manufacturer:
          type: string
        model:
          type: string
        serialNumber:
          type: string
        location:
          type: string
        installationDate:
          type: string
          format: date
        lastMaintenanceDate:
          type: string
          format: date
        nextMaintenanceDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/EquipmentStatus'
        criticalLevel:
          $ref: '#/components/schemas/CriticalLevel'
        technicalSpecs:
          type: object
          additionalProperties: true
...
    Equipment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vesselId:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: '#/components/schemas/EquipmentType'
        manufacturer:
          type: string
        model:
          type: string
        serialNumber:
          type: string
        location:
          type: string
        installationDate:
          type: string
          format: date
        lastMaintenanceDate:
          type: string
          format: date
        nextMaintenanceDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/EquipmentStatus'
        criticalLevel:
          $ref: '#/components/schemas/CriticalLevel'
        qualityScore:
          type: integer
          minimum: 0
          maximum: 100
        photos:
          type: array
          items:
            $ref: '#/components/schemas/FileUpload'
        documents:
          type: array
          items:
            $ref: '#/components/schemas/FileUpload'
        partCount:
          type: integer
        technicalSpecs:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        verifiedBy:
          type: string
          format: uuid
        verifiedAt:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
...
    FileUpload:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        url:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
        entityType:
          type: string
        entityId:
          type: string
          format: uuid
        description:
          type: string
        uploadedBy:
          type: string
          format: uuid
        uploadedAt:
          type: string
          format: date-time
...
    BatchOperationResponse:
      type: object
      properties:
        successful:
          type: integer
        failed:
          type: integer
        results:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              success:
                type: boolean
              id:
                type: string
                format: uuid
              error:
                type: string

    # List Response Schemas
...
    EquipmentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Equipment'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true

    # Enums
...
    EquipmentType:
      type: string
      enum:
        - ENGINE
        - GENERATOR
        - PUMP
        - COMPRESSOR
        - CRANE
        - WINCH
        - BOILER
        - PURIFIER
        - NAVIGATION
        - SAFETY
        - HVAC
        - ELECTRICAL
        - OTHER
    EquipmentStatus:
      type: string
      enum:
        - OPERATIONAL
        - DEGRADED
        - FAILED
        - MAINTENANCE
        - DECOMMISSIONED
    CriticalLevel:
      type: string
      enum:
        - CRITICAL
        - ESSENTIAL
        - IMPORTANT
        - STANDARD
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...